segment-extractor/requirements.txt
# Segment Extractor Dependencies

# Data processing
pandas==2.0.0
numpy==1.24.3

# HTTP requests
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Progress bars
tqdm==4.65.0

# Optional: for faster data processing
# pyarrow>=10.0.0
